refactor: move database to infra
