+++ internal/postgres/media_store.go
@@ -1,8 +1,8 @@
 package postgres
 
 import (
-	"github.com/pet-sitter/pets-next-door-api/internal/database"
 	"github.com/pet-sitter/pets-next-door-api/internal/domain/media"
+	"github.com/pet-sitter/pets-next-door-api/internal/infra/database"
 )
 
 type MediaPostgresStore struct {
